add initial support for plain text palette files
TileStache currently only supports Adobe Photoshop's .act file format
for defining a palette for png optimisation. However being a binary
format it is hard to maintain .act files or create them without
Photoshop.

Although there are many other formats in existance we could support, the
aim of this patch is to quickly and simply provide a plain text format
with the same set of features the .act format provides.